requirements.txt
numpy
pandas
scikit-learn
matplotlib
joblib
pywin32
pyyaml